feat: gera as tabelas automaticamente

+++ src/main/java/com/example/LibraryAPI/model/Livro.java
@@ -2,6 +2,7 @@ package com.example.LibraryAPI.model;
 
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Objects;
 
@@ -28,7 +29,7 @@ public class Livro {
     private GeneroLivro genero;
 
     @Column(name = "preco", precision = 18, scale = 2)
-    private Double preco;
+    private BigDecimal preco;
 
     @ManyToOne
     @JoinColumn(name = "id_autor")
